Address PR comments.
- Move NoGood.fromOldLiterals to NoGoodTest.
