+++ src/test/java/at/ac/tuwien/kr/alpha/grounder/ChoiceGrounder.java
@@ -38,7 +38,7 @@ import java.util.stream.Stream;
 
 import static at.ac.tuwien.kr.alpha.Util.entriesToMap;
 import static at.ac.tuwien.kr.alpha.Util.entry;
-import static at.ac.tuwien.kr.alpha.common.NoGood.fromOldLiterals;
+import static at.ac.tuwien.kr.alpha.common.NoGoodTest.fromOldLiterals;
 import static at.ac.tuwien.kr.alpha.common.NoGood.headFirst;
 import static java.util.Arrays.asList;
 import static java.util.Collections.singleton;
