+++ src/test/java/at/ac/tuwien/kr/alpha/solver/NaiveNoGoodStoreTest.java
@@ -6,7 +6,9 @@ import org.junit.Before;
 import org.junit.Ignore;
 import org.junit.Test;
 
-import static at.ac.tuwien.kr.alpha.common.NoGood.*;
+import static at.ac.tuwien.kr.alpha.common.NoGood.fact;
+import static at.ac.tuwien.kr.alpha.common.NoGood.headFirst;
+import static at.ac.tuwien.kr.alpha.common.NoGoodTest.fromOldLiterals;
 import static at.ac.tuwien.kr.alpha.solver.ThriceTruth.*;
 import static org.junit.Assert.*;
