+++ src/test/java/at/ac/tuwien/kr/alpha/solver/learning/GroundConflictNoGoodLearnerTest.java
@@ -5,7 +5,7 @@ import at.ac.tuwien.kr.alpha.solver.*;
 import org.junit.Ignore;
 import org.junit.Test;
 
-import static at.ac.tuwien.kr.alpha.common.NoGood.fromOldLiterals;
+import static at.ac.tuwien.kr.alpha.common.NoGoodTest.fromOldLiterals;
 import static org.junit.Assert.*;
 
 /**
